Read a paging structure #9

+++ boot/src/asm/control.rs
@@ -92,8 +92,8 @@ impl Register3 {
         cr3.into()
     }
 
-    pub fn get_page_directory_base(&self) -> u64 {
-        self.page_directory_base() << Self::PAGE_DIRECTORY_BASE_OFFSET
+    pub fn get_page_directory_base(&self) -> usize {
+        (self.page_directory_base() << Self::PAGE_DIRECTORY_BASE_OFFSET) as usize
     }
 }
